+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 # Changelog
 
+## [3.16.1](https://github.com/Zephyruso/zashboard/compare/v3.16.0...v3.16.1) (2026-07-31)
+
+
+### Bug Fixes
+
+* prevent clear button from submitting the enclosing form ([0fb5daf](https://github.com/Zephyruso/zashboard/commit/0fb5daf369f0c61d0bb41e989b6c8a08904f5e3c)), closes [#591](https://github.com/Zephyruso/zashboard/issues/591)
+* **singbox:** test individual outbound from proxy card ([b4fc4fa](https://github.com/Zephyruso/zashboard/commit/b4fc4fabe7141ee3424e3575b4ef772bf8bbfd7d))
+* **singbox:** wait for pushed latency results ([51d3327](https://github.com/Zephyruso/zashboard/commit/51d3327714478dce8a9c08c1ef5f79b96178a751))
+* update sing-box icon ([29e2516](https://github.com/Zephyruso/zashboard/commit/29e251637f4d76db7141dedb9da938bb244cd824))
+
 ## [3.16.0](https://github.com/Zephyruso/zashboard/compare/v3.15.0...v3.16.0) (2026-07-22)